Added group widget update code

+++ javahueapi/src/main/java/nl/q42/javahueapi/models/Action.java
@@ -15,7 +15,7 @@ public class Action implements Serializable {
 	
 	public int sat;
 	
-	public double[] xy;
+	public float[] xy;
 	
 	public int ct;
